fosphor/cl: Allow the use of cl_khr_local_int32_base_atomics extension
Apparently there are some cards that are neither Nvidia SM1.1, nor
OpenCL 1.1 but that do have cl_khr_local_int32_base_atomics extension.

This is enough for fosphor to implement the histogram so add the
required code to test for it and enable its use in the kernel if
required.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>