fosphor/cl: Fix tex coordinate in bicubic color mapping shader
The texture coordinates didn't match what the normal OpenGL bilinear
mapper does. It was off by half a texel.

This is needed because the color in the texture is considered to be a
sample at the center of the texel. (hence 0.5f, 0.5f offset removed)

For more details on bilinear from OpenGL :

http://hacksoflife.blogspot.be/2009/12/texture-coordinate-system-for-opengl.html

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>