gr-fosphor: Set default visibility to false
We're hidden by default.

For Qt we'll get an expose event the first time we're shown.
Needed because we might be hidden by default (like in a tab
chooser) and this would create problems if we consider ourselves
shown by default.

And for GLFW we consider ourself visible after the first reshape,
which is forced at init.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>